refactor
the stream consumer options and the stream filter options. This implementation is similar to the
golang implementation. So the idea is to have the same API ( as much as possible ) for all the amqp 1.0. clients.
Add a new class StreamFilterOptions to configure only the filtering and leave the StreamConsumerOptions for offset
and other future implementations

Signed-off-by: Gabriele Santomaggio <G.santomaggio@gmail.com>